requirements.txt
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
scikit-learn==1.3.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
joblib==1.3.2
tensorflow==2.15.0
fastapi==0.104.1
uvicorn==0.24.0.post1
